Extract Lark projection row helpers
